fixes ITP-939 and ITP-931

+++ modules/networks/signals.py
@@ -27,5 +27,6 @@ def delete_dns_records_receiver(sender, **kwargs):
 def delete_host_address_assignments_receiver(sender, **kwargs):
     if settings.TASKS_ENABLED:
         host = kwargs['instance']
-        for assignment in host.addressassignment_set.all():
-            assignment.delete()
+        if len(host.address_assignments) > 0:
+            for assignment in host.address_assignments:
+                assignment.delete()
